Tidy `RhythmEvaluator`s

+++ src/com/rian/osu/difficulty/evaluators/DroidRhythmEvaluator.kt
@@ -77,8 +77,9 @@ object DroidRhythmEvaluator {
 
             val currentRatio = 1 + 6 * min(
                 0.5,
-                sin(Math.PI / (min(prevDelta, currentDelta) / max(prevDelta, currentDelta))).pow(2.0)
+                sin(Math.PI / (min(prevDelta, currentDelta) / max(prevDelta, currentDelta))).pow(2)
             )
+
             val windowPenalty =
                 ((abs(prevDelta - currentDelta) - greatWindow * 0.6) / (greatWindow * 0.6)).coerceIn(0.0, 1.0)
 
@@ -119,8 +120,8 @@ object DroidRhythmEvaluator {
 
                     rhythmComplexitySum +=
                         sqrt(effectiveRatio * startRatio) * currentHistoricalDecay *
-                                sqrt((4 + islandSize).toDouble()) / 2 *
-                                sqrt((4 + previousIslandSize).toDouble()) / 2
+                        sqrt((4.0 + islandSize)) / 2 *
+                        sqrt((4.0 + previousIslandSize)) / 2
 
                     startRatio = effectiveRatio
                     previousIslandSize = islandSize
